+++ lib/occams-record.rb
@@ -14,6 +14,11 @@ require 'occams-record/batches/offset_limit/scoped'
 require 'occams-record/batches/offset_limit/raw_query'
 require 'occams-record/batches/cursor_helpers'
 
+require 'occams-record/binds_converter'
+require 'occams-record/binds_converter/abstract'
+require 'occams-record/binds_converter/named'
+require 'occams-record/binds_converter/positional'
+
 require 'occams-record/query'
 require 'occams-record/raw_query'